manifest: add EXTRA_PATH to executable script to allow passing a custom PATH
It seems that the flatpak CLI does not allow to pass a certain PATH when
running a flatpak.  In some cases, the user might want to do that,
still, on their own risk. This modification allows the user to pass an
EXTRA_PATH variable which will then get appended to the PATH for calling
the IDE. This allows to run executables from the host.
Of course, that will eventually break. But then it is the user who
actively opted into that breaking behaviour by passing the variable.