manifest: add EXTRA_PATH to allow passing a custom PATH
It seems that the Flatpak CLI does not allow to pass a certain PATH when running a Flatpak. In some cases, users might want to do that, still, at their own risk.

This allows the user to pass an EXTRA_PATH variable which will then get appended to the PATH for calling the IDE, allowing to run binaries/executables from the host.

Of course, that will eventually break. But then it is the user who actively opted into that breaking behavior by passing the variable.