fix: make GPG commit signing work again, remove dependencies from 74152a7
GnuPG is hard-coded to invoke `/usr/bin/pinentry`, so `/app/bin/pinentry` and
its dependencies inside the Flatpak are never used. Instead, the app is
supposed to call the gpg-agent through the socket, which then invokes
`/usr/bin/pinentry` on the host without leaking the password input inside the
Flatpak app.